pyproject.toml
[project]
name = "vinasltcalculator"
version = "0.1.0"
description = ""
authors = [
    {name = "peyro",email = "peyrovskaaa@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "fastapi (>=0.116.1,<0.117.0)",
    "alembic (>=1.16.5,<2.0.0)",
    "pydantic-settings (>=2.10.1,<3.0.0)",
    "fastapi-problem (>=0.11.5,<0.12.0)",
    "redis (>=6.4.0,<7.0.0)",
    "loguru (>=0.7.3,<0.8.0)",
    "fastapi-cache2 (>=0.2.2,<0.3.0)",
    "currencyconverter (>=0.18.9,<0.19.0)",
    "pandas (>=2.3.2,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
alembic = "^1.16.5"
aiosqlite = "^0.21.0"

